refactor(react-tinacms-inline): creates inline-image dir
